Split into two projects (public API, generator), configure new project setup to build and pack, update package references to 1.1.1
