+++ .github/workflows/nuget-publish.yml
@@ -31,5 +31,5 @@ jobs:
         run: |
           mkdir __out
           dotnet pack -c Release -o __out --no-restore
-          dotnet nuget push "./__out/EnumScribe.1.1.0.nupkg" --source "https://api.nuget.org/v3/index.json" --skip-duplicate --no-symbols true --api-key ${{secrets.NUGET_API_KEY}}
+          dotnet nuget push "./__out/EnumScribe.1.1.1.nupkg" --source "https://api.nuget.org/v3/index.json" --skip-duplicate --no-symbols true --api-key ${{secrets.NUGET_API_KEY}}
           rm __out -Recurse -Force
