+++ EnumScribe.sln
@@ -5,6 +5,11 @@ VisualStudioVersion = 17.0.31815.197
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EnumScribe", "src\EnumScribe\EnumScribe.csproj", "{DBAAE4BC-A236-4E9A-954B-C98815527E0A}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EnumScribe.Generator", "src\EnumScribe.Generator\EnumScribe.Generator.csproj", "{7A2F725F-8D12-4758-AB77-2048816E1997}"
+	ProjectSection(ProjectDependencies) = postProject
+		{DBAAE4BC-A236-4E9A-954B-C98815527E0A} = {DBAAE4BC-A236-4E9A-954B-C98815527E0A}
+	EndProjectSection
+EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EnumScribe.Samples", "samples\EnumScribe.Samples\EnumScribe.Samples.csproj", "{DAAA2A0C-B851-478C-8D45-9E8BFAF364EB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EnumScribe.Tests", "tests\EnumScribe.Tests\EnumScribe.Tests.csproj", "{B0712501-0BB4-47BE-AC3C-0D37D21D04A2}"
@@ -24,6 +29,10 @@ Global
 		{DBAAE4BC-A236-4E9A-954B-C98815527E0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DBAAE4BC-A236-4E9A-954B-C98815527E0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DBAAE4BC-A236-4E9A-954B-C98815527E0A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7A2F725F-8D12-4758-AB77-2048816E1997}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7A2F725F-8D12-4758-AB77-2048816E1997}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7A2F725F-8D12-4758-AB77-2048816E1997}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7A2F725F-8D12-4758-AB77-2048816E1997}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DAAA2A0C-B851-478C-8D45-9E8BFAF364EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DAAA2A0C-B851-478C-8D45-9E8BFAF364EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DAAA2A0C-B851-478C-8D45-9E8BFAF364EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
