+++ samples/EnumScribe.Samples/EnumScribe.Samples.csproj
@@ -7,7 +7,8 @@
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>disable</ImplicitUsings>
 		<IsPackable>false</IsPackable>
-		<NoWarn>169</NoWarn> <!-- Ignore unused private field EnumScribe.Samples.Inventory._alpacaStock -->
+		<!-- Ignore unused private field EnumScribe.Samples.Inventory._alpacaStock -->
+		<NoWarn>169</NoWarn>
 	</PropertyGroup>
 
 	<PropertyGroup Condition="'$(Configuration)'=='Debug'">
@@ -17,7 +18,8 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\..\src\EnumScribe\EnumScribe.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="true" />
+		<ProjectReference Include="..\..\src\EnumScribe\EnumScribe.csproj" />
+		<ProjectReference Include="..\..\src\EnumScribe.Generator\EnumScribe.Generator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
 	</ItemGroup>
 
 </Project>
