+++ src/EnumScribe.Generator/AnalyzerReleases.Shipped.md
