+++ src/EnumScribe.Generator/AnalyzerReleases.Unshipped.md
