+++ src/EnumScribe.Generator/Properties/launchSettings.json
