+++ src/EnumScribe/EnumScribe.csproj
@@ -4,49 +4,26 @@
 		<TargetFramework>netstandard2.0</TargetFramework>
 		<LangVersion>10.0</LangVersion>
 		<Nullable>enable</Nullable>
-		<Features>strict</Features>
-		<IsRoslynComponent>true</IsRoslynComponent>
-		<IsPackable>true</IsPackable>
+		<Deterministic>true</Deterministic>
 		<EnableNETAnalyzers>true</EnableNETAnalyzers>
+		<Features>strict</Features>
+		<!-- Packed in EnumScribe.Generator -->
+		<IsPackable>false</IsPackable>
 		<GenerateDocumentationFile>True</GenerateDocumentationFile>
-		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-		<!-- Do not include the generator as a lib dependency -->
-		<!-- Can't use this yet... need the attributes to be available. Might be able to get away with splitting the projects... -->
-		<!--<IncludeBuildOutput>false</IncludeBuildOutput>-->
 	</PropertyGroup>
 
-	<PropertyGroup Label="Nuget configuration">
-		<!--https://docs.microsoft.com/en-us/nuget/reference/msbuild-targets#pack-target-->
-		<PackageId>EnumScribe</PackageId>
-		<PackageVersion>1.1.0</PackageVersion>
-		<Authors>Tonu</Authors>
-		<Title>EnumScribe</Title>
-		<Description>Generates string properties to access enum descriptions on marked types. Primarily used to simplify binding enum description text in UI elements and avoid runtime reflection. Also allows description access via extension method for targeted enums.</Description>
-		<PackageLicenseExpression>MIT</PackageLicenseExpression>
-		<PackageProjectUrl>https://github.com/TonuFish/EnumScribe</PackageProjectUrl>
-		<PackageIcon>icon_256.png</PackageIcon>
-		<PackageIconUrl>https://github.com/TonuFish/EnumScribe/tree/master/img/icon_256.png</PackageIconUrl>
-		<PackageReadmeFile>README.md</PackageReadmeFile>
-		<PackageTags>description attribute description-attribute enum C# generator source-generator extensions extension-method</PackageTags>
-		<PackageReleaseNotes>See https://github.com/TonuFish/EnumScribe/tree/master/CHANGELOG.md for release notes.</PackageReleaseNotes>
-		<RepositoryUrl>https://github.com/TonuFish/EnumScribe.git</RepositoryUrl>
-		<RepositoryType>git</RepositoryType>
-		<RepositoryBranch>master</RepositoryBranch>
+	<PropertyGroup Label="Assembly configuration">
+		<Version>1.1.1</Version>
+		<AssemblyVersion>1.1.1.0</AssemblyVersion>
+		<FileVersion>1.1.1.0</FileVersion>
 	</PropertyGroup>
 
 	<ItemGroup Label="Package configuration">
-		<None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
-		<None Include="..\..\img\icon_256.png" Pack="true" PackagePath="\" Visible="false" />
-		<None Include="..\..\.nuget\NUGET_README.md" Pack="true" PackagePath="\" Visible="false" />
-		<AdditionalFiles Include="AnalyzerReleases.Shipped.md" Pack="true" />
-		<AdditionalFiles Include="Internal\AnalyzerReleases.Unshipped.md" />
-		<AdditionalFiles Include="PublicAPI.Shipped.txt" Pack="true" />
-		<AdditionalFiles Include="Internal\PublicAPI.Unshipped.txt" />
+		<AdditionalFiles Include="PublicAPI.Shipped.txt" />
+		<AdditionalFiles Include="PublicAPI.Unshipped.txt" />
 	</ItemGroup>
 
 	<ItemGroup Label="Private references">
-		<PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.3.3" PrivateAssets="all" />
-		<PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.0.1" PrivateAssets="all" />
 		<PackageReference Include="Microsoft.CodeAnalysis.PublicApiAnalyzers" Version="3.3.3" PrivateAssets="all">
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
