--- src/EnumScribe/Properties/AssemblyInfo.cs
@@ -1,1 +1,0 @@
-﻿[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("EnumScribe.Tests")]
