+++ src/EnumScribe/PublicAPI.Unshipped.txt
